requirements.txt
Pillow
easygui